requirements.txt
# PyTorch GPU con CUDA 11.8
torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# Dependencias restantes
mediapipe==0.10.10
pandas
h5py
jiwer
numpy==1.26.4
opencv-python==4.8.1.78
git+https://github.com/princeton-vl/RAFT.git

# Backend API
fastapi
uvicorn[standard]
grpcio
protobuf

